Update and rename AStarVisualizer.jsx to AStarVisualizer.tsx
